+++ apollo-composition/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- [#590](https://github.com/apollographql/federation-rs/pull/590) Fix
+  deserialization of `GraphQLError` nodes.
+
 ## 0.1.4
 
 - Update to `apollo-federation` 2.0.0-preview.0